fix compilation errors on 64bits architectures
 * sizeof(unsigned int) != sizeof(size_t)
 * INT_MAX != LONG_MAX

+++ lasso/saml-2.0/provider.c
@@ -153,7 +153,7 @@ xsdUnsignedShortParse(xmlChar *value, int *out) {
 
 	errno = 0;
 	l = strtol((char*)value, NULL, 10);
-	if (((l == LONG_MIN || l == LONG_MAX) && errno == ERANGE) ||
+	if (((l == INT_MIN || l == INT_MAX) && errno == ERANGE) ||
 			errno == EINVAL || l < 0 || l >= 65535) {
 		return FALSE;
 	}
